# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kebelet Open API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Review List",
                "description": "Review List here",
                "operationId": "reviewList",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Paginate size",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Paginate page",
                        "required": false
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "filter by user_id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Review Create",
                "description": "Review Create here",
                "operationId": "reviewCreate",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "toilet_id",
                                    "rating",
                                    "description"
                                ],
                                "properties": {
                                    "toilet_id": {
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "type": "integer"
                                    },
                                    "total_toilet": {
                                        "type": "string"
                                    },
                                    "cleanness": {
                                        "type": "string"
                                    },
                                    "facility": {
                                        "type": "string"
                                    },
                                    "crowded": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review successfully created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reviews/upload": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Review Images",
                "description": "Upload multiple review images",
                "operationId": "Upload Review Image",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "reviews_id",
                                    "img[]"
                                ],
                                "properties": {
                                    "reviews_id": {
                                        "type": "integer"
                                    },
                                    "img[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "file",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Files Successfully uploaded",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reviews/{id}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Review Retrieve",
                "description": "Review Retrieve here",
                "operationId": "reviewRetrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "summary": "Review Delete",
                "description": "PrReviewoduct Delete here",
                "operationId": "reviewDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to delete the Review."
                    },
                    "404": {
                        "description": "Review not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}